+++ Sources/SwiftDisc/Internal/JSONValue.swift
@@ -1,0 +1,24 @@
+import Foundation
+
+public enum JSONValue: Encodable {
+    case string(String)
+    case number(Double)
+    case int(Int)
+    case bool(Bool)
+    case object([String: JSONValue])
+    case array([JSONValue])
+    case null
+
+    public func encode(to encoder: Encoder) throws {
+        var container = encoder.singleValueContainer()
+        switch self {
+        case .string(let s): try container.encode(s)
+        case .number(let n): try container.encode(n)
+        case .int(let i): try container.encode(i)
+        case .bool(let b): try container.encode(b)
+        case .object(let o): try container.encode(o)
+        case .array(let a): try container.encode(a)
+        case .null: try container.encodeNil()
+        }
+    }
+}